Rename pricePreDiscount to basePrice
