+++ WooCommerce/Classes/ViewRelated/Orders/Order Creation/EditableOrderViewModel.swift
@@ -698,7 +698,7 @@ final class EditableOrderViewModel: ObservableObject {
                                                                   imageURL: variation.imageURL,
                                                                   name: item.name,
                                                                   sku: variation.sku,
-                                                                  price: item.pricePreDiscount.stringValue,
+                                                                  price: item.basePrice.stringValue,
                                                                   pricedIndividually: pricedIndividually,
                                                                   discount: passingDiscountValue,
                                                                   productTypeDescription: ProductType.variable.description,
@@ -745,7 +745,7 @@ final class EditableOrderViewModel: ObservableObject {
                                                                   imageURL: product.imageURL,
                                                                   name: product.name,
                                                                   sku: product.sku,
-                                                                  price: item.pricePreDiscount.stringValue,
+                                                                  price: item.basePrice.stringValue,
                                                                   pricedIndividually: pricedIndividually,
                                                                   discount: passingDiscountValue,
                                                                   productTypeDescription: product.productType.description,
